feature: bark bark

+++ apps/imagine-api/src/photo/photo.resolver.ts
@@ -2,11 +2,13 @@ import {In} from 'typeorm';
 import {PhotoModel} from './photo.model';
 import {UserModel} from '../user/user.model';
 import {RoomModel} from '../room/room.model';
-import {Inject, forwardRef} from '@nestjs/common';
+import {UserEntity} from '../database/user.entity';
 import {PhotoEntity} from '../database/photo.entity';
+import {GetUser} from '../session/get-user.decorator';
 import {RoomRepository} from '../database/room.repository';
 import {UserRepository} from '../database/user.repository';
 import {PhotoRepository} from '../database/photo.repository';
+import {Inject, UnauthorizedException, forwardRef} from '@nestjs/common';
 import {PhotoFilterManyInput, PhotoFilterOneInput} from './photo.input';
 import {
   Args,
@@ -67,7 +69,12 @@ export class PhotoResolver {
   }
 
   @Mutation(() => Boolean)
-  async photoDelete(@Args('id') id: number) {
+  async photoDelete(@Args('id') id: number, @GetUser() session: UserEntity) {
+    const matchingPhoto = await this.photoRepo.findOneOrFail({id: id});
+    const userOwnsPhoto = matchingPhoto.userID === session.id!;
+    if (!userOwnsPhoto) {
+      throw new UnauthorizedException();
+    }
     await this.photoRepo.delete({id});
     return true;
   }
